Break routes packages into files by function
